% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_help.R
\name{checkSampleSizeEnv}
\alias{checkSampleSizeEnv}
\title{Return objects from SampleSizeEnv environment}
\usage{
checkSampleSizeEnv(intervalMillis, session = NULL, check, name)
}
\arguments{
\item{intervalMillis}{Approximate number of milliseconds to wait between
checking the variable check.}

\item{session}{The user session to associate this file reader with, or NULL
if none. If non-null, the reader will automatically stop when the session
ends.}

\item{check}{character indicating what to check. If this has changed the
object referred to by check is fectched}

\item{name}{The object to fetch from the SampleSizeEnv}
}
\value{
reactive function with the object stored in SampleSizeEnv with name: name
}
\description{
Return objects from SampleSizeEnv environment
}
