% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_avatar.R
\name{create_avatar}
\alias{create_avatar}
\title{Creates the avatar for the person who logged in.}
\usage{
create_avatar(
  user,
  col = sample(nncol$company[-1], length(user), replace = length(user) >=
    length(nncol$company)),
  edge = col,
  edge.size = 0.3,
  height = 128,
  text.size = height/9,
  type = c("circle", "square"),
  save_png = FALSE,
  overwrite = FALSE,
  file = paste0(paste(tolower(user), collapse = "_"), ".png")
)
}
\arguments{
\item{user}{The user name the avatar is based upon}

\item{col}{The background color.}

\item{edge}{color of the edge in \code{type} equal to \code{circle}}

\item{edge.size}{size of the edge for \code{type} equal to \code{circle}}

\item{height}{The height of the image if saved as png}

\item{text.size}{Size of text used for}

\item{type}{should the avatars be squares or circles}

\item{save_png}{should a png be save}

\item{overwrite}{Should a new avatar be created if it allready exists}

\item{file}{If saved the path to the file name.}
}
\value{
Saves a png file that can be used in the comment section
}
\description{
Creates the avatar for the person who logged in.
}
\examples{
create_avatar("sffl")
online <- c("cdn", "sffl")

users <- c("cdn", "agtr", "abiu", "sffl", "kfsm")

online_col <- (2:3)[users \%in\% online + 1]
create_avatar(c(users), edge = online_col, text.size = 128 / 12)
create_avatar(c(users), edge = online_col, text.size = 128 / 12, edge.size = 0.8)

}
