% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalHR.R
\name{criticalHR}
\alias{criticalHR}
\title{Calculate the critical hazard ratio}
\usage{
criticalHR(n_events, z_alpha)
}
\arguments{
\item{n_events}{Number of events}

\item{z_alpha}{One sided significance level}
}
\value{
Returns the critical hazard ratio
}
\description{
Calculate the critical hazard ratio
}
\examples{
exp(-2*1*sqrt(1/40))
}
