% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_col.R
\name{get_prop_col}
\alias{get_prop_col}
\title{Get proportion color}
\usage{
get_prop_col(prop)
}
\arguments{
\item{prop}{The proportion}
}
\value{
Returns the correct proportion color
}
\description{
Get proportion color
}
\examples{
0.8
}
