% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss_estimates.R
\name{get_ss_estimates}
\alias{get_ss_estimates}
\title{Get sample size estimates}
\usage{
get_ss_estimates(
  lambdaC,
  eta,
  minfup,
  T_max,
  sided,
  ratio,
  hr,
  alpha,
  beta,
  interim,
  timing,
  sfu
)
}
\arguments{
\item{lambdaC}{Yearly placebo event rate (percent) for patients (Hazard rate C)}

\item{eta}{Equal dropout hazard rate for both groups}

\item{minfup}{Minimum follow up}

\item{T_max}{Maximum study duration}

\item{sided}{One or two sided test}

\item{ratio}{Randomization ratio between placebo and treatment group}

\item{hr}{Hazard rate}

\item{alpha}{The significance level}

\item{beta}{1-beta is the power}

\item{interim}{Should the analysis include interim or not}

\item{timing}{The timing of interim}

\item{sfu}{Spending function when including interim}
}
\value{
Returns the calculated sample size estimates
}
\description{
Core function for computing number of events and subjects
}
