% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumSubjRecruit.R
\name{getNumSubjRecruit}
\alias{getNumSubjRecruit}
\title{Get number of subjects to be recruited, when considering screening failure risk and 
dropout risk.}
\usage{
getNumSubjRecruit(
  scr_fail_risk,
  dropout_risk_trt,
  dropout_risk_ctrl,
  cal_sample_size
)
}
\arguments{
\item{scr_fail_risk}{Screening failure risk}

\item{dropout_risk_trt}{Dropout risk in treatment group}

\item{dropout_risk_ctrl}{Dropout risk in control group}

\item{cal_sample_size}{Calculated sample size}
}
\value{
Returns the number of subjects to be recruited to obtain the calculated sample size
}
\description{
Get number of subjects to be recruited, when considering screening failure risk and 
dropout risk.
}
\examples{
list
}
