% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleSizeBioequi.R
\name{getSampleSizeBioequi}
\alias{getSampleSizeBioequi}
\title{Get sample size estimate for bioequivalence trials}
\usage{
getSampleSizeBioequi(
  targetpower,
  theta0_AUC,
  theta0_CMAX,
  CV_AUC,
  CV_CMAX,
  studyDesign
)
}
\arguments{
\item{targetpower}{The minimum power to be obtained}

\item{theta0_AUC}{Assumed AUC ratio}

\item{theta0_CMAX}{Assumed CMAX ratio}

\item{CV_AUC}{Coefficient of variations (percent) of the AUC ratio}

\item{CV_CMAX}{Coefficient of variations (percent) of the CMAX ratio}

\item{studyDesign}{Study design, eg. parallel and cross-over}
}
\value{
Returns the calculated sample size estimates
}
\description{
Core function for computing number of events and subjects
}
