% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSampleSizereview.R
\name{hDrive}
\alias{hDrive}
\title{Returns the path to the h drive}
\usage{
hDrive(path = "", ...)
}
\arguments{
\item{path}{A \code{character} specifying a path to a subfolder.}

\item{...}{More paths (passed to \code{\link{paste}}).}
}
\value{
Windows systems returns \code{"//whale.novo.dk/<initial>/"} whereas
  non-windows systems returns \code{file.path("/z/falcon/home/<initial>/"}
  followed by any given paths.
}
\description{
Based on the OS, this function returns the path on the user specific ("H")
drive.
}
