% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\title{Create the header for EOT style document}
\usage{
header(
  x,
  title = c("Title 1", "Title 2", "Title 3", "Title 4"),
  status = c("Draft", "Review", "Final"),
  version = "0.1",
  date = format(Sys.Date(), "\%d \%b \%Y")
)
}
\arguments{
\item{x}{The document to change}

\item{title}{Character of length 4 giving the title for the header}

\item{status}{character defining the status of the document}

\item{version}{character defining the version, e.g. "0.1"}

\item{date}{The date to include in the header}
}
\value{
The document with the header altered as specified
}
\description{
Create the header for EOT style document
}
