% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_help.R
\name{loginInfo}
\alias{loginInfo}
\title{Login info for a current run}
\usage{
loginInfo(check = FALSE)
}
\arguments{
\item{check}{Should the function return last update time or the actual login info}
}
\value{
An object of class \code{tibble} with info on user logins
}
\description{
Login info for a current run
}
