% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDownloadBttn.R
\name{myDownloadBttn}
\alias{myDownloadBttn}
\title{Download button}
\usage{
myDownloadBttn(
  outputId,
  label = "Download",
  style = "bordered",
  color = "default",
  size = "md",
  block = FALSE,
  no_outline = FALSE,
  icon_ = icon("download")
)
}
\arguments{
\item{outputId}{The id of the output}

\item{label}{To be downloaded}

\item{style}{The style}

\item{color}{The color}

\item{size}{The size}

\item{block}{True or false}

\item{no_outline}{True or false}

\item{icon_}{The icon for the button}
}
\value{
Downloads the output
}
\description{
Download button
}
