% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_col.R
\docType{data}
\name{nncol}
\alias{nncol}
\title{Vector containing company colours}
\format{
An object of class \code{list} of length 1.
}
\usage{
nncol
}
\description{
A vector containing company colours
}
\details{
This vector is used by the shiny application
\code{\link{runSampleSizereview}} and the function \code{\link{create_avatar}}.
}
\examples{
# show the colour
nncol$company
}
\keyword{datasets}
