% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathInput.R
\name{pathInput}
\alias{pathInput}
\title{This function generates the client-side HTML for a path input}
\usage{
pathInput(inputId, label, value = "", width = NULL)
}
\arguments{
\item{inputId}{The id of the input}

\item{label}{The label}

\item{value}{The initiating value}

\item{width}{The width of the input}
}
\description{
This function generates the client-side HTML for a path input
}
