% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSampleSizereview.R
\name{pDrive}
\alias{pDrive}
\title{Returns the path to the project drive}
\usage{
pDrive(path = "", ...)
}
\arguments{
\item{path}{A \code{character} specifying a path to a subfolder.}

\item{...}{More paths (passed to \code{\link{paste}}).}
}
\value{
Windows systems returns \code{"//whale.novo.dk/projstat/"} and
  non-windows systems returns \code{"/projstat/"} followed by any giving
  paths.
}
\description{
Based on the OS, this function returns the path on the project drive.
}
