% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.R
\name{pwr}
\alias{pwr}
\title{Calculate the power}
\usage{
pwr(end_type, n, sig.level, power, trt, ctrl, sd, alternative, determine)
}
\arguments{
\item{end_type}{The endpoint test type}

\item{n}{Number of observations}

\item{sig.level}{Significance level}

\item{power}{Power of test}

\item{trt}{Mean in treatment group}

\item{ctrl}{Men in control group}

\item{sd}{Standard deviation in both groups}

\item{alternative}{Alternative hypothesis}

\item{determine}{Set to NA}
}
\value{
Returns the power calculation
}
\description{
Calculate the power
}
