% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSampleSizereview.R
\name{readTable}
\alias{readTable}
\title{Read a file into R}
\usage{
readTable(
  path,
  sep = "auto",
  sep2 = "\\\\|",
  which_sep2 = NULL,
  sep3 = "\\032",
  which_sep3 = NULL,
  ...
)
}
\arguments{
\item{path}{path to the file}

\item{sep}{The separator between columns. Defaults to the character in the
set \code{[,\\t |;:]} that separates the sample of rows into the most
number of lines with the same number of fields. Use \code{NULL} or
\code{""} to specify no separator; i.e. each line a single character column
like \code{base::readLines} does.}

\item{sep2}{separator to be used within columns}

\item{which_sep2}{The columns to use sep2 on, either as integer with column
numbers or character with column names. If Null all character columns are
used.}

\item{sep3}{Third level separator}

\item{which_sep3}{columns to use the third level sep on.}

\item{...}{Arguments passed on to fread}
}
\value{
the loaded data as tibble
}
\description{
Read a file into R
}
