% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSampleSizereview.R
\name{runSampleSizereview}
\alias{runSampleSizereview}
\title{Run the pk pd review tool}
\usage{
runSampleSizereview(
  trial = NULL,
  username = unname(Sys.info()["user"]),
  workspace = c("local", "server", "shinyServer"),
  verbose = FALSE,
  port = trial,
  access_data_path = file.path(getwd(), "persistent_data"),
  data_storage = file.path(access_data_path, "trial_data"),
  admins = username,
  ...
)
}
\arguments{
\item{trial}{The four digit trial to review}

\item{username}{Username to initialise the app without logging in}

\item{workspace}{is the app run locally or at server}

\item{verbose}{Should text be written in the console to describe what is
currently running. Used for debugging.}

\item{port}{The to initialise the app on.}

\item{access_data_path}{Path to store application data}

\item{data_storage}{\code{character} with either path to folder for data
storage or \code{"trial_specific"} for path to be defined in the UI.}

\item{admins}{\code{character} indicating user that admins}

\item{...}{Additional arguments passed on to \cite{shiny}{runApp}}
}
\value{
Runs the server app with the specified settings
}
\description{
Run the pk pd review tool
}
\examples{
\dontrun{
runSampleSizereview()
}
}
