% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_environment.R
\docType{data}
\name{samplesize_environment}
\alias{samplesize_environment}
\alias{nnr_environment}
\alias{nnr_env}
\title{The NNSampleSize-package enviroment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
samplesize_environment
}
\description{
An environment for various NNSampleSize package objects for session caching and
package parameters. Used internally by the NNSampleSize package.
}
\details{
This environment is used by the shiny application \code{\link{runSampleSizereview}}.

Using an environment for this is useful as we do not
accidentally create copies of the environment.
}
\keyword{internal}
