% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_environment.R
\name{SampleSizeEnv}
\alias{SampleSizeEnv}
\title{Working with the SampleSize environment}
\usage{
SampleSizeEnv(
  x,
  value,
  remove = FALSE,
  clear_environment = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{The variable name to either get, set or remove from the SampleSize
environement. For Getting and setting variables x is a variable name, given as
a character string. No coercion is done, and the first element of a
character vector of length greater than one will be used, with a warning.
For remove it can be a vector of characters of the names that should be removed.}

\item{value}{A value to be assigned to x}

\item{remove}{logical indicating if the objects given in x should be removed.}

\item{clear_environment}{Should the environment be cleared completely}

\item{verbose}{Should the function be verbose.}

\item{...}{Aditional arguments passed on to lower level functions.}
}
\value{
Gets, sets, or removes the elements supplied in x from the environment
}
\description{
Working with the SampleSize environment
}
\examples{

# Assign the value 2 to object a
SampleSizeEnv("a", 2, verbose = TRUE)

# Get the value assigned to object a
SampleSizeEnv("a", verbose = TRUE)

# Remove object "a" from the environment
SampleSizeEnv("a", remove = TRUE, verbose = TRUE)

# When the objects are not found
SampleSizeEnv("a", verbose = TRUE)
SampleSizeEnv("a", remove = TRUE, verbose = TRUE)

# assign and clear the environement
SampleSizeEnv("a", 2)
SampleSizeEnv("b", 3)

# What does the environment contain
SampleSizeEnv()

# Clear the environment
SampleSizeEnv(clear_environment = TRUE, verbose = TRUE)
SampleSizeEnv()
}
