% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_help.R
\name{shinyInput}
\alias{shinyInput}
\title{Create multiple similar shiny inputs for a table}
\usage{
shinyInput(FUN, n, id, ...)
}
\arguments{
\item{FUN}{\code{function} The shiny input function to use}

\item{n}{\code{numeric} giving The number of shiny inputs needed}

\item{id}{Id \code{character} with an id}

\item{...}{Additional arguments passed on to function \code{FUN}}
}
\value{
\code{character} of length n containing the shiny inputs with id
  given as  paste0(id, i) where i is the row number
}
\description{
Create multiple similar shiny inputs for a table
}
\examples{
if (require("shiny"))
  shinyInput(actionButton, 4, "a_good_id_", "go")
}
