% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathInput.R
\name{timeInput}
\alias{timeInput}
\title{Time input}
\usage{
timeInput(inputId, label, value = "")
}
\arguments{
\item{inputId}{The id of the input}

\item{label}{The label}

\item{value}{The initiating value}
}
\description{
This function generates the client-side HTML for a URL input
}
