% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_help.R
\name{updateLoginInfo}
\alias{updateLoginInfo}
\title{Update log in info}
\usage{
updateLoginInfo(session.id, user, login = FALSE, logout = FALSE)
}
\arguments{
\item{session.id}{A unique session id}

\item{user}{the user}

\item{login}{time of login}

\item{logout}{time of logout}
}
\value{
The function is used for the side effects. It returns TRUE
}
\description{
Update log in info
}
