% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathInput.R
\name{updatePathInput}
\alias{updatePathInput}
\title{Send an update message to a path input on the client.}
\usage{
updatePathInput(session, inputId, label = NULL, value = NULL)
}
\arguments{
\item{session}{The session to update}

\item{inputId}{The id of the path}

\item{label}{The label}

\item{value}{The initiating path}
}
\description{
This update message can change the value and/or label.
}
