% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSampleSizereview.R
\name{writeTable}
\alias{writeTable}
\title{Write a data table to a file}
\usage{
writeTable(
  x,
  path,
  sep = ";",
  sep2 = c("", "|", ""),
  sep3 = "\\032",
  which_sep3 = NULL,
  ...
)
}
\arguments{
\item{x}{The tibble to save}

\item{path}{The path to the file}

\item{sep}{The separator between columns}

\item{sep2}{The separator within columns}

\item{sep3}{Third level separator}

\item{which_sep3}{columns to use the third level sep on.}

\item{...}{Arguments passed on to fwrite}
}
\value{
Saves x to the file supplied in path
}
\description{
Write a data table to a file
}
